+++ bauh/api/paths.py
@@ -10,3 +10,4 @@ USER_THEMES_PATH = '{}/.local/share/bauh/themes'.format(str(Path.home()))
 DESKTOP_ENTRIES_DIR = '{}/.local/share/applications'.format(str(Path.home()))
 TEMP_DIR = f'/tmp/{__app_name__}@{getuser()}'
 LOGS_DIR = f'{TEMP_DIR}/logs'
+AUTOSTART_DIR = f'/etc/xdg/autostart' if user.is_root() else f'{Path.home()}/.config/autostart'